+++ src/import-statements/css-scss.ts
@@ -8,10 +8,10 @@ import { ImportStyle } from '../model';
  * @param {string} relativePath Relative path of dragged file and active text editor.
  * @returns Import statement string
  */
-export function getCSSImport(relativePath: string): vscode.SnippetString {
+export function cssImportStatement(relativePath: string): vscode.SnippetString {
 
-  let configValue = vscode.workspace.getConfiguration('importStatements.styleSheet').get('cssImportStyle');
-      configValue = importStyle.css.find((config: ImportStyle<number>) => config.description === configValue).value;
+  let configValue = vscode.workspace.getConfiguration('auto-import.importStatement.styleSheet').get('cssImportStyle');
+      configValue = importStyle.css.find((config: ImportStyle) => config.description === configValue).value;
 
   switch (configValue as number) {
     case 0:  return new vscode.SnippetString(`@import '${relativePath}';`);
@@ -26,10 +26,10 @@ export function getCSSImport(relativePath: string): vscode.SnippetString {
  * @param {string} relativePath Relative path of dragged file and active text editor.
  * @returns Import statement string
  */
- export function getCSSImageImport(relativePath: string): vscode.SnippetString {
+ export function cssImageImportStatement(relativePath: string): vscode.SnippetString {
 
-  let configValue = vscode.workspace.getConfiguration('importStatements.styleSheet').get('cssImageImportStyle');
-      configValue = importStyle.cssImage.find((config: ImportStyle<number>) => config.description === configValue).value;
+  let configValue = vscode.workspace.getConfiguration('auto-import.importStatement.styleSheet').get('cssImageImportStyle');
+      configValue = importStyle.cssImage.find((config: ImportStyle) => config.description === configValue).value;
 
   switch (configValue as number) {
     case 0:  return new vscode.SnippetString(`url('${relativePath}')`);
@@ -43,10 +43,12 @@ export function getCSSImport(relativePath: string): vscode.SnippetString {
  * @param {string} relativePath Relative path of dragged file and active text editor.
  * @returns Import statement string
  */
-export function getSCSSImport(relativePath: string): vscode.SnippetString {
+export function scssImportStatement(relativePath: string): vscode.SnippetString {
 
-  let configValue = vscode.workspace.getConfiguration('importStatements.styleSheet').get('scssImportStyle');
-      configValue = importStyle.scss.find((config: ImportStyle<number>) => config.description === configValue).value;
+  relativePath = parsePartialFile(relativePath);
+
+  let configValue = vscode.workspace.getConfiguration('auto-import.importStatement.styleSheet').get('scssImportStyle');
+      configValue = importStyle.scss.find((config: ImportStyle) => config.description === configValue).value;
 
   switch (configValue as number) {
     case 0:  return new vscode.SnippetString(`@import '${relativePath}';`);
@@ -63,14 +65,21 @@ export function getSCSSImport(relativePath: string): vscode.SnippetString {
  * @param {string} relativePath Relative path of dragged file and active text editor.
  * @returns Import statement string
  */
-export function getSCSSImageImport(relativePath: string): vscode.SnippetString {
+export function scssImageImportStatement(relativePath: string): vscode.SnippetString {
 
-  let configValue = vscode.workspace.getConfiguration('importStatements.styleSheet').get('scssImageImportStyle');
-      configValue = importStyle.scssImage.find((config: ImportStyle<number>) => config.description === configValue).value;
+  let configValue = vscode.workspace.getConfiguration('auto-import.importStatement.styleSheet').get('scssImageImportStyle');
+      configValue = importStyle.scssImage.find((config: ImportStyle) => config.description === configValue).value;
 
   switch (configValue as number) {
     case 0:  return new vscode.SnippetString(`url('${relativePath}')`);
     default: return new vscode.SnippetString(`url('${relativePath}')`);
   }
 
+}
+
+function parsePartialFile(relativePath: string): string {
+  const arr = relativePath.split('/');
+  const lastElemIndex = arr.length - 1;
+  arr[lastElemIndex].startsWith('_') && (arr[lastElemIndex] = arr[lastElemIndex].substring(1));
+  return arr.join('/');
 }