+++ src/import-statements/index.ts
@@ -1,0 +1,9 @@
+import * as htmlMarkdown from './html-markdown';
+import * as cssScss from './css-scss';
+import * as javascriptTypescript from './javascript-typescript';
+
+export const importStatement = {
+  ...htmlMarkdown,
+  ...cssScss,
+  ...javascriptTypescript
+};