+++ src/import-texts/index.ts
@@ -1,3 +1,0 @@
-export * from './html-markdown';
-export * from './css-scss';
-export * from './javascript-typescript';