+++ src/utilities/index.ts
@@ -2,4 +2,4 @@ export * from './file-extension';
 export * from './import-type';
 export * from './notify';
 export * from './relative-path';
-export * from './import-text';
+export * from './import-statement-snippet';